requirements.txt
tavily-python
langchain
python-dotenv
psycopg2
streamlit
google-generativeai